Split timeline provider
